[GridFragmentation] Basic support of break-before/break-after on item.
This implements basic support for break-before/break-after. At the
moment it only support the top level break-before/break-after value (not
the nested values, until we set up the non-fragmented constraint space
correctly).

The break-before/break-after values are collected at the grid row level
when we calculate the grid (stored in a vector).

Currently this also doesn't (yet) search for the row with the highest
appeal.

Bug: 614667
Change-Id: I0299fa5d7b04106641903607c84fe45470f8a27c