shapedetection: Remove reachable NOTREACHED()
When handling the various forms of a V8ImageBitmapSource there was a
NOTREACHED() statement on the 3 unsupported source types, even though
the detect() function also had code to throw a NotSupportedError
DOMException. Since these cases are in fact reachable when those
ImageBufferSource types are passed in the NOTREACHED() is inappropriate.

It has been removed and tests are added to exercise this case.

Bug: 1275599
Change-Id: Ie4c4367aea9f75ec60fd3c10cd297184881ecff7